src/main/java/com/recipe/spring5recipeapp/converters/UnitOfMeasureCommandToUnitOfMeasure.java
package com.recipe.spring5recipeapp.converters;

import com.recipe.spring5recipeapp.commands.UnitOfMeasureCommand;
import com.recipe.spring5recipeapp.domain.UnitOfMeasure;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

//wire it up as spring bean
@Component
public class UnitOfMeasureCommandToUnitOfMeasure implements Converter<UnitOfMeasureCommand, UnitOfMeasure> {

    //spring doesn't make it thread safe, so we added @Synchronized
    @Synchronized
    @Nullable
    @Override
    public UnitOfMeasure convert(UnitOfMeasureCommand source) {
        if (source == null){
            return null;
        }
        final UnitOfMeasure uom = new UnitOfMeasure();
        uom.setId(source.getId());
        uom.setDescription(source.getDescription());
        return uom;
    }
}
