src/test/java/com/recipe/spring5recipeapp/converters/UnitOfMeasureCommandToUnitOfMeasureTest.java
package com.recipe.spring5recipeapp.converters;

import com.recipe.spring5recipeapp.commands.UnitOfMeasureCommand;
import com.recipe.spring5recipeapp.domain.UnitOfMeasure;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class UnitOfMeasureCommandToUnitOfMeasureTest {

    private static final String description = "description";
    private static final Long id = 1L;

    UnitOfMeasureCommandToUnitOfMeasure converter;

    @BeforeEach
    void setUp() {
        converter = new UnitOfMeasureCommandToUnitOfMeasure();
    }

    @Test
    void testNullParameter(){
        assertNull(converter.convert(null));
    }

    @Test
    void testEmptyObject(){
        assertNotNull(converter.convert(new UnitOfMeasureCommand()));
    }

    @Test
    void convert() {
        //given
        UnitOfMeasureCommand command = new UnitOfMeasureCommand();
        command.setId(id);
        command.setDescription(description);

        //when
        UnitOfMeasure uom = converter.convert(command);

        //then
        assertNotNull(uom);
        assertEquals(id, uom.getId());
        assertEquals(description, uom.getDescription());
    }
}